requirements.txt
fastapi
uvicorn
pymongo
kafka-python
python-dotenv
openai
textblob
